+++ datafusion/core/src/physical_optimizer/mod.rs
@@ -24,7 +24,6 @@
 
 pub mod enforce_distribution;
 pub mod enforce_sorting;
-pub mod join_selection;
 pub mod optimizer;
 pub mod projection_pushdown;
 pub mod replace_with_order_preserving_variants;
