+++ datafusion/physical-optimizer/Cargo.toml
@@ -41,6 +41,7 @@ datafusion-execution = { workspace = true }
 datafusion-expr-common = { workspace = true, default-features = true }
 datafusion-physical-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
+futures = { workspace = true }
 itertools = { workspace = true }
 log = { workspace = true }
 recursive = { workspace = true, optional = true }
@@ -50,3 +51,4 @@ datafusion-expr = { workspace = true }
 datafusion-functions-aggregate = { workspace = true }
 datafusion-functions-nested = { workspace = true }
 tokio = { workspace = true }
+rstest = { workspace = true }
